Create $faker in _inject instead of _before
We often use _before, so if _before is used in a test with this trait,
we'd have to remember to call the parent method, and calling it creates
an extra line of non-obvious code. We don't use Codeception dependency
injection at all, so I put the creation of Faker\Generator here.